requirements.txt
mcp>=0.9.0
openai>=1.0.0
discord.py>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0